Added credit to the image and also added this version of the page to GitHub.
This was my first attempt to making a multi-page website. I realised that I didn't have it on GitHub, so I thought i'd add it. I also wanted to add credit to the image of the croissant that I have used in this page.
This is a fictional multi-page for croissants, as part of SheCodes Responsive.